+++ hack/tests/e2e-scaleupdown-previous.sh
@@ -33,9 +33,15 @@ T="$(mktemp -d)"
 start_monitoring $T/src/github.com/openshift/openshift-azure/_data/containerservice.yaml
 
 git clone -b "$1" https://github.com/openshift/openshift-azure.git $T/src/github.com/openshift/openshift-azure
+ln -sf "$PWD/secrets" "$T/src/github.com/openshift/openshift-azure"
 (
+    set +x
+    export AZURE_MASTER_CLIENT_ID=$AZURE_LEGACY_MASTER_CLIENT_ID
+    export AZURE_MASTER_CLIENT_SECRET=$AZURE_LEGACY_MASTER_CLIENT_SECRET
+    export AZURE_WORKER_CLIENT_ID=$AZURE_LEGACY_WORKER_CLIENT_ID
+    export AZURE_WORKER_CLIENT_SECRET=$AZURE_LEGACY_WORKER_CLIENT_SECRET
+    set -x
     cd "$T/src/github.com/openshift/openshift-azure"
-    setup_secrets
     GOPATH="$T" make create
 )