+++ hack/vmimage.sh
@@ -26,11 +26,12 @@ fi
 
 export IMAGE_RESOURCEGROUP="${IMAGE_RESOURCEGROUP:-images}"
 export IMAGE_RESOURCENAME="${IMAGE_RESOURCENAME:-rhel7-3.11-$(TZ=Etc/UTC date +%Y%m%d%H%M)}"
+export IMAGE_STORAGEACCOUNT="${IMAGE_STORAGEACCOUNT:-openshiftimages}"
 
 go generate ./...
-go run -ldflags "-X main.gitCommit=$COMMIT" ./cmd/vmimage -imageResourceGroup "$IMAGE_RESOURCEGROUP" -image "$IMAGE_RESOURCENAME"
+go run -ldflags "-X main.gitCommit=$COMMIT" ./cmd/vmimage -imageResourceGroup "$IMAGE_RESOURCEGROUP" -image "$IMAGE_RESOURCENAME" -imageStorageAccount "$IMAGE_STORAGEACCOUNT"
 
-export AZURE_REGION=eastus
+export AZURE_REGIONS=eastus
 export RESOURCEGROUP="${IMAGE_RESOURCENAME//./}-e2e"
 
 make create