+++ test/e2e/specs/realrp/vnetpeering.go
@@ -3,7 +3,6 @@ package realrp
 import (
 	"context"
 	"fmt"
-	"os"
 	"time"
 
 	. "github.com/onsi/ginkgo"
@@ -11,6 +10,7 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2018-06-01/network"
 
+	v20190430 "github.com/openshift/openshift-azure/pkg/api/2019-04-30"
 	"github.com/openshift/openshift-azure/pkg/fakerp/client"
 	"github.com/openshift/openshift-azure/test/clients/azure"
 	tlog "github.com/openshift/openshift-azure/test/util/log"
@@ -28,15 +28,11 @@ var _ = Describe("Peer Vnet tests [Vnet][Real][LongRunning]", func() {
 		cli, err = azure.NewClientFromEnvironment(false)
 		Expect(err).NotTo(HaveOccurred())
 
-		// setting dummy aad
-		os.Setenv("AZURE_AAD_CLIENT_ID", os.Getenv("AZURE_CLIENT_ID"))
-		os.Setenv("AZURE_AAD_CLIENT_SECRET", os.Getenv("AZURE_CLIENT_SECRET"))
-
-		cfg, err = client.NewConfig(tlog.GetTestLogger(), true)
+		cfg, err = client.NewConfig(tlog.GetTestLogger())
 		Expect(err).NotTo(HaveOccurred())
 
 		// create a new resource group
-		_, err = client.CreateResourceGroup(cfg)
+		err = client.EnsureResourceGroup(cfg)
 		Expect(err).ToNot(HaveOccurred())
 	})
 
@@ -85,7 +81,8 @@ var _ = Describe("Peer Vnet tests [Vnet][Real][LongRunning]", func() {
 		Expect(len(*vnet.VirtualNetworkPeerings)).To(Equal(0))
 
 		// load cluster config
-		config, err := client.GenerateManifest("../../test/manifests/realrp/create.yaml")
+		var config v20190430.OpenShiftManagedCluster
+		err = client.GenerateManifest(cfg, "../../test/manifests/realrp/create.yaml", &config)
 		Expect(err).ToNot(HaveOccurred())
 
 		// Set pre-created peer vnetid in cluster config
@@ -93,7 +90,7 @@ var _ = Describe("Peer Vnet tests [Vnet][Real][LongRunning]", func() {
 
 		// create a cluster with the peerVnet
 		By("creating an OSA cluster")
-		_, err = cli.OpenShiftManagedClusters.CreateOrUpdateAndWait(ctx, cfg.ResourceGroup, cfg.ResourceGroup, *config)
+		_, err = cli.OpenShiftManagedClusters.CreateOrUpdateAndWait(ctx, cfg.ResourceGroup, cfg.ResourceGroup, config)
 		Expect(err).NotTo(HaveOccurred())
 
 		By("ensuring the OSA cluster vnet is peered with the custom vnet")